move budget structures into a separate file
