+++ lib/collection/benches/batch_search_bench.rs
@@ -12,8 +12,8 @@ use collection::optimizers_builder::OptimizersConfig;
 use collection::save_on_disk::SaveOnDisk;
 use collection::shards::local_shard::LocalShard;
 use collection::shards::shard_trait::ShardOperation;
+use common::budget::ResourceBudget;
 use common::counter::hardware_accumulator::HwMeasurementAcc;
-use common::cpu::ResourceBudget;
 use criterion::{criterion_group, criterion_main, Criterion};
 use rand::rng;
 use segment::data_types::vectors::{only_default_vector, VectorStructInternal};
