+++ lib/collection/src/collection_manager/optimizers/vacuum_optimizer.rs
@@ -212,8 +212,8 @@ mod tests {
     use std::sync::atomic::AtomicBool;
     use std::sync::Arc;
 
+    use common::budget::ResourcePermit;
     use common::counter::hardware_counter::HardwareCounterCell;
-    use common::cpu::ResourcePermit;
     use itertools::Itertools;
     use parking_lot::RwLock;
     use segment::entry::entry_point::SegmentEntry;
