+++ lib/collection/src/operations/snapshot_storage_ops.rs
@@ -2,7 +2,7 @@ use std::fs::File;
 use std::io::{BufReader, Read};
 use std::path::{Path, PathBuf};
 
-use common::cpu::ResourceBudget;
+use common::budget::ResourceBudget;
 use futures::StreamExt;
 use object_store::WriteMultipart;
 use tokio::io::AsyncWriteExt;
