+++ lib/collection/src/shards/replica_set/update.rs
@@ -651,7 +651,7 @@ mod tests {
     use std::num::NonZeroU32;
     use std::sync::Arc;
 
-    use common::cpu::ResourceBudget;
+    use common::budget::ResourceBudget;
     use segment::types::Distance;
     use tempfile::{Builder, TempDir};
     use tokio::runtime::Handle;
