+++ lib/collection/src/shards/shard_holder/mod.rs
@@ -8,7 +8,7 @@ use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::time::Duration;
 
-use common::cpu::ResourceBudget;
+use common::budget::ResourceBudget;
 use common::tar_ext::BuilderExt;
 use futures::{stream, Future, StreamExt, TryStreamExt as _};
 use itertools::Itertools;
