+++ lib/collection/src/tests/points_dedup.rs
@@ -3,8 +3,8 @@ use std::num::NonZeroU32;
 use std::sync::Arc;
 
 use api::rest::OrderByInterface;
+use common::budget::ResourceBudget;
 use common::counter::hardware_accumulator::HwMeasurementAcc;
-use common::cpu::ResourceBudget;
 use rand::{rng, Rng};
 use segment::data_types::vectors::NamedVectorStruct;
 use segment::types::{
