+++ lib/common/common/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod bitpacking;
 pub mod bitpacking_links;
 pub mod bitpacking_ordered;
+pub mod budget;
 pub mod counter;
 pub mod cpu;
 pub mod defaults;
