+++ lib/segment/tests/integration/byte_storage_quantization_test.rs
@@ -3,7 +3,7 @@ use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 
 use atomic_refcell::AtomicRefCell;
-use common::cpu::ResourcePermit;
+use common::budget::ResourcePermit;
 use common::types::ScoredPointOffset;
 use itertools::Itertools;
 use rand::prelude::StdRng;
