+++ lib/segment/tests/integration/exact_search_test.rs
@@ -2,8 +2,8 @@ use std::collections::HashMap;
 use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 
+use common::budget::ResourcePermit;
 use common::counter::hardware_counter::HardwareCounterCell;
-use common::cpu::ResourcePermit;
 use common::types::PointOffsetType;
 use itertools::Itertools;
 use rand::{rng, Rng};
