+++ lib/segment/tests/integration/multivector_filtrable_hnsw_test.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 
-use common::cpu::ResourcePermit;
+use common::budget::ResourcePermit;
 use common::types::TelemetryDetail;
 use itertools::Itertools;
 use rand::prelude::StdRng;
