+++ lib/storage/src/content_manager/toc/mod.rs
@@ -27,8 +27,9 @@ use collection::shards::replica_set::{AbortShardTransfer, ReplicaState};
 use collection::shards::shard::{PeerId, ShardId};
 use collection::shards::{replica_set, CollectionId};
 use collection::telemetry::CollectionTelemetry;
+use common::budget::ResourceBudget;
 use common::counter::hardware_accumulator::HwSharedDrain;
-use common::cpu::{get_num_cpus, ResourceBudget};
+use common::cpu::get_num_cpus;
 use common::types::TelemetryDetail;
 use dashmap::DashMap;
 use tokio::runtime::Runtime;
