+++ lib/storage/tests/integration/alias_tests.rs
@@ -5,7 +5,7 @@ use collection::operations::vector_params_builder::VectorParamsBuilder;
 use collection::operations::verification::new_unchecked_verification_pass;
 use collection::optimizers_builder::OptimizersConfig;
 use collection::shards::channel_service::ChannelService;
-use common::cpu::ResourceBudget;
+use common::budget::ResourceBudget;
 use memory::madvise;
 use segment::types::Distance;
 use storage::content_manager::collection_meta_ops::{
