+++ src/main.rs
@@ -17,7 +17,8 @@ use std::thread;
 use std::thread::JoinHandle;
 use std::time::Duration;
 
-use ::common::cpu::{get_cpu_budget, get_io_budget, ResourceBudget};
+use ::common::budget::{get_io_budget, ResourceBudget};
+use ::common::cpu::get_cpu_budget;
 use ::tonic::transport::Uri;
 use api::grpc::transport_channel_pool::TransportChannelPool;
 use clap::Parser;
